demo-mcp-server/pyproject.toml
[project]
name = "demo-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["mcp[cli]>=1.9.4", "python-dotenv>=1.1.0"]
